requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
flask
joblib
comet-ml 
dvc
dvc-gs
flask
pillow